packages/openai-financial-helper/pyproject.toml
[tool.poetry]
name = "openai-financial-helper"
version = "0.0.1"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
langchain-core = ">=0.1.5"
langchain-openai = ">=0.0.1"
langchain_experimental = ">=0.0.55"
pandas = "*"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.4"
fastapi = "^0.104.0"
sse-starlette = "^1.6.5"

[tool.langserve]
export_module = "openai_financial_helper"
export_attr = "chain"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
